Some minor changes

+++ src/gui/MapView.hpp
@@ -39,16 +39,16 @@ protected:
 	//! The scene
 	QGraphicsScene scene;
 	//! The world map image
-    QPixmap pix;
-    QGraphicsPixmapItem* pixItem;
-    //! The pointeur image
-    QPixmap pointeurPix;
-    //! The seleted position
-    QPointF pointeurPos;
-    //! The list of all the cities in the map
-    QList<City> cities;
-    //! Set to true just after we select a new position
-    bool justSelected;
+	QPixmap pix;
+	QGraphicsPixmapItem* pixItem;
+	//! The pointeur image
+	QPixmap pointeurPix;
+	//! The seleted position
+	QPointF pointeurPos;
+	//! The list of all the cities in the map
+	QList<City> cities;
+	//! Set to true just after we select a new position
+	bool justSelected;
 public:
 	//! Get the scale at wich we are seeing the map
 	float getScale() const {return matrix().m11();}
@@ -61,16 +61,16 @@ public:
 	//! Select a city
 	void select(const City* city);
 	//! Select a position
-	void select(float longitude, float latitude);
+	void select(double longitude, double latitude);
 	//! Highlight a city (0) if unhighlighted
 	void highlightCity(const City* city);
 signals:
 	//! Signal emitted when we select a new location
 	//! If no city is selected we set city to the empty string
-	void positionSelected(float longitude, float latitude, QString city);
+	void positionSelected(double longitude, double latitude, QString city);
 	//! Signal emitted when we move over a city
 	//! When we move out we emmit the same signal with city set to the empty string
-	void positionHighlighted(float longitude, float latitude, QString city);
+	void positionHighlighted(double longitude, double latitude, QString city);
 protected:
 	//! Add all the cities into the scene
 	void populate(const QString& filename = "data/cities_Earth.fab");
